more JS codes on showNext function
Learned more about the event and how to set event target, previous version failed as wrong understanding on how to select a group of elements with the same class names, if set a variable = the className, this variable is only the first element with the same name in this class group. Have to use an upper element that contains this elements group and target the specific elements, in this case, the option buttons.